// src/api/generated/model/creativeTaskSubmissionEventDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Title
 * Description
 * OpenAPI spec version: 1.0
 */
import type { CreativeTaskSubmissionEventDtoType } from './creativeTaskSubmissionEventDtoType';
import type { CreativeTaskSubmissionEventDtoActorType } from './creativeTaskSubmissionEventDtoActorType';
import type { CreativeTaskSubmissionEventDtoPayload } from './creativeTaskSubmissionEventDtoPayload';

export interface CreativeTaskSubmissionEventDto {
  /** ID события */
  id: string;
  /** Момент события */
  createdAt: string;
  /** Системное имя события */
  type: CreativeTaskSubmissionEventDtoType;
  /** Кто совершил действие */
  actorType: CreativeTaskSubmissionEventDtoActorType;
  /**
   * Снимок данных на момент события (комментарий, статусы, ссылка и т.п.)
   * @nullable
   */
  payload?: CreativeTaskSubmissionEventDtoPayload;
}