// src/api/generated/model/creativeTaskSubmissionEventDtoType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Title
 * Description
 * OpenAPI spec version: 1.0
 */

/**
 * Системное имя события
 */
export type CreativeTaskSubmissionEventDtoType = typeof CreativeTaskSubmissionEventDtoType[keyof typeof CreativeTaskSubmissionEventDtoType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CreativeTaskSubmissionEventDtoType = {
  created: 'created',
  materials_submitted: 'materials_submitted',
  materials_approved: 'materials_approved',
  materials_rejected: 'materials_rejected',
  materials_updated: 'materials_updated',
  publication_url_updated: 'publication_url_updated',
  publication_reported: 'publication_reported',
  publication_approved: 'publication_approved',
  publication_rejected: 'publication_rejected',
  comment_updated: 'comment_updated',
} as const;